Update patch set 1
Patch Set 1:

Chris, thank you for moving the spec to Doctor Git repo. There's still much work to be done until final draft is ready for submission to OpenStack, but this is the way to go!

This .rst differs from the one in the etherpad. E.g. blueprint URL to OpenStack contains typos leading to 404 page in Launchpad, and you included the entire text from etherpad even though the blueprint ends with the Reference section (what's below should be kept out the .rst IMO). I would also rename the file name to the name the blueprint is registed in OpenStack (update-server-state-immeinletdiately in this case) and add it in a subdirectory "nova" as it relates to OpenStack Nova.

Keeping neutron vote for now. Would like to hear also from other members first.

Patch-set: 1