Update patch set 6
Patch Set 6: Code-Review-1

(33 comments)

Hi, 

Thanks for the answers for my previous review. I managed to read the second half of the document and put questions and comments inline.

Is there a way to generate html from the rst in this repo? For patchset 4 I was able to use a tool rst2html.py but for patchset 6 it fails for multiple reasons.

Patch-set: 6
Reviewer: Gerrit User 713 <713@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1