Removing configurability of phys-virtual mapping
The affected part of the document is unclear and missleading.

The relation of physical resource to virtualized resource are
created by VIM and in some cases the process that selects physical
resource for a virtual resource request are not 100% configurable.
For example during VM scheduling in nova if multiple hosts can
fulfill the request and the hosts has the same weight then nova
simply select the first from the list which can be considered as
a random choice.

This patch is based on the related unresolved comment in
https://gerrit.opnfv.org/gerrit/#/c/304

JIRA: DOCTOR-4
Signed-off-by: Balazs Gibizer <balazs.gibizer@ericsson.com>