Add reference to service-status-notification bp
To make easier to cover the Maintenance Notification gap
a new bp was proposed to have AMQP notification for administrative
nova service status change. As soon as this bp gets implemented
Doctor (via Ceilometer) can get notification if the status of
nova service (e.g. nova-compute) is changed administratively.
Changing the nova-compute service status to down means that
nova will not use the compute node any more for new instances
so this is very much in line with putting a node in maintenance
node.

Signed-off-by: Balazs Gibizer <balazs.gibizer@ericsson.com>
Change-Id: I71a1621eb177ae6dd7da1a09285da68d398a30e8