Update patch set 1
Patch Set 1:

Hi Balazs,

Thank you for kicking this amendment.

I think the sentence is not right - the intention was "The relation from failure in physical resource to failure in virtualized resource shall be configurable".

The VIM may have failure selection and aggregation in three points: a) phy-->phy, b) phy-->vir and c) vir-->vir. It might be nice to have this paragraph as description of b), or edit related contents including this sentence.

Patch-set: 1
Label: Verified=0