Update patch set 2
Patch Set 2:

One will not have to follow any changes manually because the blueprint was merged upstream and thus it's not going to suffer further changes. We in Doctor project and other requirement projects decided at some point in time to merge the changes back from upstream to OPNFV once they are approved. Happy to revisit this agreement, though.

Ryota, the note in design_docs/notification-alarm-evaluator.rst is meant to inform community that we halted development in OPNFV and sent it upstream. You're suggesting to add a note but now saying something like "Blueprint was submitted and merged in OpenStack: <link>"? Either that and deleting all the specification text or merge changes from OpenStack to OPNFV as I proposed in this patch.

Patch-set: 2
Label: Verified=0