Update patch set 2
Patch Set 2:

> One will not have to follow any changes manually because the
 > blueprint was merged upstream and thus it's not going to suffer
 > further changes. We in Doctor project and other requirement
 > projects decided at some point in time to merge the changes back
 > from upstream to OPNFV once they are approved. Happy to revisit
 > this agreement, though.
Unfortunately it is not a valid assumption that a spec once merged upstream will not change in the future. There is at least two scenarion in my head when a new patchset might be created
for an already approved and merges spec:

* during the implementation of the spec it turns out that the spec is faulty and need to be
changed. In this case a new patch needs to be updated with the new design to keep the implementation and the spec in synch
* implementation doesn't get merged in a given release therefore the spec needs to be reproposed to the next release and because the baseline is changed in the mean time the spec needs to be updated as well.

Cheers,
Gibi

Patch-set: 2