Update patch set 2
Patch Set 2:

Hi,

I agree with Balazs that we should link the approved blueprint from the wiki and/or the deliverable document. I don't see a value in duplicating the material, especially that there are cases, when the approved spec is still not the final one and evolves further.

Best Regards,
Ildiko

Patch-set: 2
Label: Code-Review=0