Adding fencing as general feature
This patch is based on a related unresolved comment in
https://gerrit.opnfv.org/gerrit/#/c/304

At an earlier stage, we had identified fencing as a gap in Nova, but
this gap was removed. Who remembers why we made this decision? Is the
feature already sufficiently implemented in Nova? The related BP seems
obsoleted.

JIRA: DOCTOR-10

Co-Authored-By: Gerald Kunzmann <kunzmann@docomolab-euro.com>
Signed-off-by: Carlos Goncalves <carlos.goncalves@neclab.eu>
Change-Id: Id31ac1552a8f1e3506c5e4d533416611d6b95217