Update patch set 1
Patch Set 1:

Thank you, Bertrand!

I have several questions here:

(1) Is the "1. Maintenance Request" intended to sync operation in http/rest level? If so, it's may cause some issues such as timeout (depending on configs). I suppose it can be async by using notification mechanism for response or asking admins to check status of resources on the host in another operation.

(3) What is the "Timeout" of "1. Maintenance Request"? Timeout of the entire maintenance request, or time out of execution of instructed actions? Is this the same one used in step 6 of Fig. 3.5?

Patch-set: 1