Update patch set 1
Patch Set 1:

Hi Ryota,

Yes, I will add the blank lines.

As for the LICENSE file, from my last discussion with Chris my understanding was that the images should use the same license as the rst files, which is the Creative Commons. If the final decision is that the license used in the root of the repo is enough I will remove.

Best Regards,
Ildiko

Patch-set: 1
Label: Verified=0