Update patch set 2
Patch Set 2: Code-Review+2

Ryota, is this a temporary fix for now or you plan to keep it this way? I think keeping this way could potentially bring us problems in the future, e.g. if the installers change IP addresses, etc. I am okay with changing as you propose as a workaround for Brahmaputra.

Patch-set: 2
Label: Code-Review=+2