Update patch set 2
Patch Set 2:

Carlos, this is not temporary one but can be kept at least for Brahmaptura. 'INSTALLER_IP' is set by jenkins depend on pod/installer, so we won't have problem while we use this test script in OPNFV CI pipeline. I can see a potential issuse that is a connection problem between functest contatiner, in which this script runs, and compute host. At this moment, I expect we can reach compute host from functest container. And it is yes when deploying with apex. That's also the reason why I put monitor module in functest container.


Gerald, I'm not 100% sure, but hostname of compute host should be registered in DNS server which is referred by the installer VM. We can get IP address in the installer VM, so that is not a big problem.

Patch-set: 2