test: respect 'CI_DEBUG' set by functest
In current bash script for doctor test, '-x' option is enabled and all
executing commands will be printed in the console.

Functest has a configuration to enable/disable this '-x' option. It is
environment variable 'CI_DEBUG'.

This patch makes doctor test script to set '-x' option according to
'CI_DEBUG'. If it's not set, '-x' option will be enabled as it was
enabled before this change.

JIRA: FUNCTEST-212

Change-Id: I9fa5036aad7a0f5c88378337d946439a6ee6108c
Signed-off-by: Ryota MIBU <r-mibu@cq.jp.nec.com>