Update patch set 1
Patch Set 1:

Yujun, in a single node deployment Aodh and the consumer app will be collocated in the same machine and hence 127.0.0.1/localhost is the same for both components. In a multi-node deployment, Aodh may be placed in a different machine than where the consumer app is running at. Thus, Aodh will send alarm to http://127.0.0.1/ instead of http://consumer_ip_address/

Patch-set: 1