Update patch set 1
Patch Set 1:

> Yujun, in a single node deployment Aodh and the consumer app will
 > be collocated in the same machine and hence 127.0.0.1/localhost is
 > the same for both components. In a multi-node deployment, Aodh may
 > be placed in a different machine than where the consumer app is
 > running at. Thus, Aodh will send alarm to http://127.0.0.1/ instead
 > of http://consumer_ip_address/

For a multinode deployment, it can not be assumed that the Aodh can access the docker IP address. It is in the only assumption that doctor_container is deployed in the jumphost for the default route can pass to jumpserver. Docker daemon is only valid in localhost itself(including the ip addresses allocated), it is not for network access. 
I suggest the doctor container should expose the relevant 12345/12346 port to the host using -p parameters.

Patch-set: 1
Label: Code-Review=0