Update patch set 1
Patch Set 1:

Hi all,
Sorry for reply with a delay.
Yes,it's a problem how to get the correct IP.
In my test Env, the docker which is started by functest has one IP exclude 127.0.0.1, so the command can be works.
One solution: For `local`, use 127.0.0.1, for `apex` and `fuel`, use command `ip` to get the first ip because it has the only one ip in docker.
The second soultion: suggest functest to expose a public IP and port to use.
Any other suggestions or soultions?

Patch-set: 1
Label: Verified=0