Unset default COMPUTE_USER at the beginning of tests
Each installer has different compute users and we set a default for each
in get_compute_host_info. Having COMPUTE_USER set at the beginning as is
and not passing in a value will default to 'none' and thus overriding
the default user of the installer. For example, COMPUTE_USER would be
set to 'none' instead of 'heat-admin' for Apex installer.

Change-Id: Ia60a88f68c52604bac6b7eef18e97e11328f1a6e
Signed-off-by: Carlos Goncalves <carlos.goncalves@neclab.eu>