use --os-* options instead of env param
Parameters defined in another shell cannot be used in the main shell.
This leads bug in funcsion which execute another shell to performe
openstack command as doctor test user.

This patch makes the test script use --os-* options to performe as
doctor test user, so that we don't have to execute new shells in
the test script.

Change-Id: If148f00acae919a926bb5e9f2af8dd433a4110f5
Signed-off-by: Ryota MIBU <r-mibu@cq.jp.nec.com>