Update patch set 5
Patch Set 5:

> In general I very much appreciate you contributing with Fuel
 > support!
 > 
 > At this exact point in time I am a bit reluctant in having this
 > patch merged until we have all tests with green light in local and
 > apex installers.

Yes,i agree. And we still need fuel to support the ceilometer configuration about notifier `alarm.all` topic .

BTW,the test run sucessfully in my devstack env.
But the time interval between the instance that an event is detected by monitor and the consumer received the notification is large than 1 second. It always between 4 and 8.
+ calculate_notification_time
++ grep 'doctor monitor detected at' monitor.log
++ awk '{print $5}'
+ detected=1469585626.39
++ grep 'doctor consumer notified at' consumer.log
++ awk '{print $5}'
+ notified=1469585634.0
+ echo '1469585634.0 1469585626.39'
+ awk '{d = $1 - $2; if (d < 1 && d > 0) print d " OK"; else print d " NG"}'
7.61 NG

Patch-set: 5
Label: Verified=0