Update patch set 2
Patch Set 2:

@Daniel, reading deploy.sh, I can say there is mis-configuration or bug in fuel. Fuel deploys aodh-listener, but doesn't set notification topic correctly.
We have a plan to integrate congress and vitrage later on. Do you or anyone have a plan to develop those plugins?
Our motivation is to push codes into upstream (nova, fuel, etc.) as much as we can, although putting code in doctor repo for the first step.

Patch-set: 2