Update patch set 2
Patch Set 2:

> quick question, do you intend that we'll put other configs into
 > this plugin (e.g. congress, vitrage support)?
 > 
 > If yes, I'm OK to name this as doctor plugin. But, if no, I suggest
 > you to use 'even-alarm' plugin instead.

@ryota

I think this is example for Doctor project to implement plugin in Fuel. This name can also be used for vitrage or others. In the next release I think maybe these changes in the plugin itself will updated by Fuel itself. This pulgin can still be used with content updated.

Patch-set: 2