Update patch set 3
Patch Set 3: Code-Review-1

TBH I'm not any fond of having an installer plugin modifying configurations of the base platform or of other components not managed by the plugin itself. Such way may clash with future platform configuration changes and/or other 3rd party plugins.

In particular to this plugin as it is being proposed right now, the only thing it does is config modifications to Ceilometer and that's something that should be done elsewhere in either Fuel base or any Ceilometer-related plugin/component Fuel uses to build the system.

Patch-set: 3
Label: Code-Review=-1