Update patch set 3
Patch Set 3:

> TBH I'm not any fond of having an installer plugin modifying
 > configurations of the base platform or of other components not
 > managed by the plugin itself. Such way may clash with future
 > platform configuration changes and/or other 3rd party plugins.
 > 
 > In particular to this plugin as it is being proposed right now, the
 > only thing it does is config modifications to Ceilometer and that's
 > something that should be done elsewhere in either Fuel base or any
 > Ceilometer-related plugin/component Fuel uses to build the system.

Agree with Szilard. Time is not enough for us to persuade Fuel to update this changes. the response from Fuel is it may affect other scenarios, and we can not ensure none scenario is affected. This is a generic plugin for Doctor to integrate with Fuel installer.

Patch-set: 3