Update patch set 3
Patch Set 3:

> How can you ensure non-Doctor scenarios are not affected by this
 > plugin either? If you activate it, it may or may not break same
 > way.

I think the concerns from fuel team are almost the same. See discussion in https://jira.opnfv.org/browse/FUEL-159. If this plugin will affect other scenario, what will the difference if we integrated it directly in upstream? Providing this plugin in doctor is an opt-in for doctor project, while integrating in upstream is an opt-out for all the other scenarios.

So, IMHO, if this modification is dedicated for doctor scenario, it is reasonable to be plugged in by doctor project. Please correct me if anything misunderstood.

Patch-set: 3