Update patch set 6
Patch Set 6:

> Sorry but I still stand with my earlier argument. Worse than
 > conflicting with other scenarios at testing stage is conflicting
 > later at production stag once one enables this plugin and later
 > finds out that previously running use cases started to fail.

Hi, Carlos. Still not understanding the difference between modifying this configuration in ceilometer and in doctor plugin. Shouldn't the result be the same? Or are you suggesting that we add the option of enabling all alarm notification in ceilometer plugin so people won't be surprised that ceilometer configuration is modified by a non-ceilometer plugin?

Patch-set: 6