Update patch set 3
Patch Set 3:

> (7 comments)
 > 
 > What are the modifications in tests/clean.py for? It seems applying
 > the same logic as the old code but the description is very
 > confusing.
 > 
 > OK for modification in tests/run.sh

The modification is for restore the VMS state to active. The clean.py is not called in run.sh before. Before we call the `nova force down` to set host is down, we set all the vms' states in the host as error. So after the test runs finished, restore the states.

Patch-set: 3