Update patch set 9
Patch Set 9:

Hi @Gerald, I really have no idea that why the document is still failed. I find a similar case of "..include::" in opnfvdocs. It has the same usage with our change.

https://gerrit.opnfv.org/gerrit/gitweb?p=opnfvdocs.git;a=blob;f=docs/configurationguide/configuration.options.render.rst;h=f8f16ff547f04d39593aba42ea3e139b8bee4f9e;hb=refs/heads/stable/colorado

How about rename the "figure-p1" to resolve the error that the figure numbering is not shown correct?

Patch-set: 9
Label: Verified=0