Update patch set 1
Patch Set 1:

This script will be kicked by script in releng. I thought we can have the main part of testing script in doctor repo rather than releng, so that we can fix it frequently. Anyhow, maybe we can put the script into releng first and see how it is convenient for us.

Patch-set: 1