Update patch set 1
Patch Set 1: -Code-Review

Ok, that's fine then (confirmed in 2 out of 3 PODs).

Could you include in the documentation a paragraph noting that Fuel this is only valid under a functest environment and that cloud operators cannot expect to have Doctor supported in Fuel out of the box at this point in time unless they do modifications to Nova and Ceilometer (as we are doing right now in our test script)?

Patch-set: 1
Label: Code-Review=0