Update patch set 1
Patch Set 1:

awk will return with an exit status, either 0 (OK) or 1 (not OK). run.sh is configured to exit when any command returns with an exist status different than 0. In this case, if the notification time is >0 and <1 seconds it will return 0 (OK), otherwise return 1 (NOK) and aborts.

Patch-set: 1