Update patch set 2
Patch Set 2:

I don't like looping through all compute nodes if we have a single CLI command that can do it for us. Once we're in the position of changing the role from 'admin' to '_member_' we just need to run the command as admin user instead to get the right privileges to see the "OS-EXT-SRV-ATTR:host" attribute.

Patch-set: 2