Run the tests with existing image
When ran in functest on arm, the doctor suite fails because the x86
image is hardocded in the source files. This change adds the option to
run the doctor suite with an existing image as opposed to downloading
and creating it.

The image name is taken from an enviroment variables (if it is set).
When ran with functest, the env variable is set in doctor.py file.

Change-Id: Ib6ce56e650ebe8b29618a9a31ec990f9f0895f54
Signed-off-by: cristinapauna <cristina.pauna@enea.com>
(cherry picked from commit 34e593954e048a86c90dd851de6a84c7329a937b)