Update patch set 2
Patch Set 2:

Regarding the test image, we are using a 3-part cirros-aarch64 image which is built by the Fuel installer, it's called TestVM. It's 3-part because guest EFI boot is currently not working for aarch64. This image also needs some extra boot args so it's rather complicated to modify doctor to do all that. But it is in fact a cirros image.

Patch-set: 2