don't use performance as functest result criteria
Test runs failed because of the notification time is larger than 1S in some PODs.
Find out that the time(1.983/2.26) almost used in from the inspect received the
event to nova-api begin to handle the reset_state by checking the log. So don't
use performance as functest result criteria currently as it is a sample inspector,
not topology aware.

add the time in inspector when received the failure event.

Change-Id: I2526060e60d79fbf1c2f81a494df393c653f50a8
Signed-off-by: dongwenjuan <dong.wenjuan@zte.com.cn>