Update patch set 3
Patch Set 3:

The inspector we are using ('sample') could be more lightweight than it is. Proof of that is that in other scenarios (e.g. apex) and PODs it runs just fine. We must not start disabling things from our tests in order to get green light just because the tests don't pass.

If the problem with fuel cannot be troubleshoot, we just drop it for Colorado. In any case, Fuel will not support Doctor out of the box for the reasons we've already discussed.

Patch-set: 3