Update patch set 3
Patch Set 3:

> Maybe to get the list of VMs on each host when initialize and then
 > update that list according to "instance.update" notification? But
 > this is still sample Inspector. Maybe not worth doing in sample,
 > but have that kind of thing for real one.
Yes? in the real one, we need to get all VMs when initialize and update the VMs. But in the sample inspector, first get the VMs in the host can improve the performance. It's something like topology aware, know which VMs in the host. I tested it in my ZTE-POD3, it passed. The time is less than 1S.

Patch-set: 3