Update patch set 3
Patch Set 3:

Agreed with the technical comments.

 > Please, what's the rush anyway? Fuel will not support Doctor in
 > Colorado release regardless this patch gets accepted.

Why blocking ourselves while we can still make advance?

We did not mean to rush it into Colorado if the resolution can not be ready on time. But making stable branch does not mean we have to freeze development in master, does it?

Patch-set: 3