Update patch set 4
Patch Set 4:

> (3 comments)
 > 
 > How and in which environment are you running the performance
 > profiling?

This is sample, not actual measurement. I should put some notes in the spec.

 > 
 > Would it be possible to have an automated profiling check in place
 > publishing the results under Qtip umbrella? A graphical
 > visualization of the data would be nice to have.

This spec draft shows the goal we want to achieve. Next we need to investigate the best way to implement it. There are several solutions to be tried out.

Before that, I hope to have the target agreed within the team.

Patch-set: 4