Update patch set 7
Patch Set 7:

IMHO, the Introduction is still confusing and the purpose of this patch is not clear. It sounds like a description of a set of tasks around performance profiling. Based on it, which changes do you plan to implement for Doctor, e.g. update the test scripts?
Also, it might be placed in the wrong path, as I don't see this as an own "scenario".
At this stage, we could collect such TODO's as Jira tickets. Once it is implemented, we could use this patch with the Figures e.g. as a subclause of the test case description in http://artifacts.opnfv.org/doctor/review/23963/scenarios/index.html

Patch-set: 7