Update patch set 7
Patch Set 7: Code-Review-1

I don't see the value this adds to be merged as part of the tests code as is at this moment. It's polluting unnecessarily the code base. I encourage you to continue submitting patch sets to Gerrit as way of validating your work, but PoC/demo code should not be part of master branch.

Yujun, I understand Doctor is Qtip's pet project and I really appreciate your efforts in profiling performance. My concern is that the way it seems to be taken will clog every project's test codes, not only Doctor's.

Patch-set: 7