Update patch set 7
Patch Set 7:

> I don't see the value this adds to be merged as part of the tests
 > code as is at this moment. It's polluting unnecessarily the code
 > base. I encourage you to continue submitting patch sets to Gerrit
 > as way of validating your work, but PoC/demo code should not be
 > part of master branch.

This is OK for me since different project will have different strategy to master. Keeping if validated is what I am seeking. We can talk about when to merge when it is mature.

 > Yujun, I understand Doctor is Qtip's pet project and I really
 > appreciate your efforts in profiling performance. My concern is
 > that the way it seems to be taken will clog every project's test
 > codes, not only Doctor's.

It seems a misunderstanding, it is not QTIP pushing code to projects, but project which requires performance benchmarking to integrate QTIP. As I said, it is a service and it is the project's choice to use it or not. As a PoC, I don't expect everything to be perfect, so please understand the possible pollution.

In one word, it is project's decision on whether, when and how to merge a new feature, and I'm fine with keeping it in gerrit patch until mature enough to merge.

Patch-set: 7