Update patch set 7
Patch Set 7: -Code-Review

Thanks a lot for the clarification, Yujun. I apologize if I sounded harsh on my previous comment.

We/Doctors are interested in performance profiling and support this activity overall. Please continue the prototyping, and let us know once you think you've reached to a state that code is mature enough to be merged.

Patch-set: 7
Label: Code-Review=0