Update patch set 13
Patch Set 13: Code-Review+2

Yujun, thank you for your hard work on this and addressing all reviewers' comments! Overall I'm now supportive of merging. I'm not a big fan of the filename "profiler-poc.py" or its placement but I'm not going to block it either. Perhaps a directory tests/profiling/ or tests/profiler/ would be a good place to put it in, specially if there are more related files related coming soon.

Julien, feel free to start submitting patch sets to Gerrit of the framework you're working on. That will help the team get acquainted to it and provide feedback. Depending on its complexity, the team may ask you to discuss it in the weekly meetings.

Patch-set: 13
Reviewer: Gerrit User 132 <132@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+2