Update patch set 4
Patch Set 4:

Hi Carlos,

valid point on when to reference to master or release files.
I am a bit concerned that having to update the links per release is an extra effort. Also, patches that are being cherry picked to the release may require changing the links contained in it.

I would like to get some more feedback from others on how to best address this.

Patch-set: 4