Update patch set 8
Patch Set 8:

From doctor-verify-fuel-sample-master#40:

+ openstack role remove _member_ --user doctor --project doctor
No user with a name or ID of 'doctor' exists.
+ openstack role remove admin --user admin --project doctor
No project with a name or ID of 'doctor' exists.
+ openstack project delete doctor
Failed to delete project with name or ID 'doctor': No project with a name or ID of 'doctor' exists.
1 of 1 projects failed to delete.
+ openstack user delete doctor
Failed to delete user with name or ID 'doctor': No user with a name or ID of 'doctor' exists.
1 of 1 users failed to delete.

Patch-set: 8
Label: Verified=0