Update patch set 2
Patch Set 2:

Hi Ryota,

on the right place for "scenarios": is the scenario description we have really the same as they expect in docs/release/installation?
My understanding was that in the release documentation you have a general description of the available scenarios.

Sofia, Ryota, can you please check?

Patch-set: 2