Fix run.sh bug for dynamic interfaces
The run.sh extracts the name of interface using awk
with 7th value in "ip a" command but in case of DHCP
enabled interface this value is "dynamic". So the fix
takes the last value of the output which is the
correct interface name.
EDIT: Removed the redundant line to get 5th element.

JIRA: DOCTOR-109

Change-Id: I0775d90ef219387e5b68abd4fd98d38d2104f128
Signed-off-by: Umar Farooq <umar.farooq@neclab.eu>