Add Option for Additional Interface monitoring for local installer
By default, run.sh resolves the hostname to get the IP for
monitoring. This is in most cases the admin IP of the monitored
node. To monitor additional interfaces or management network,
an additional function is added. This can be extended so that the
test figures out the network configuration through user help or automatically.

JIRA: DOCTOR-112

Change-Id: I16fff526fa3a2dd4feaedbaa7c7acbd436f83ac5
Signed-off-by: Umar Farooq <umar.farooq@neclab.eu>