Add doctor package
It also updates test-requirements.txt to conform with OpenStack
stable/ocata and renames it requirements.txt.

tests.main:main is considered as entry point and run.sh as script.
run.sh could be renamed to avoid possible conflicts in functest.

tox.ini is updated to pin the right versions and build files are
filtered.

The next dependencies can be safely removed:
  - paramiko
  - scp
  - python-openstackclient
  - python-ceilometerclient
  - virtualenv

Change-Id: I2091cfe4751640bf83d77e1fdd713af1e31fa89d
Signed-off-by: Cédric Ollivier <cedric.ollivier@orange.com>