[WIP] Test port data plane status on Sample Inspector
Exercise the Neutron port data plane status API on the Sample Inspector.

Setting the data_plane_status field of Neutron Port objects to 'DOWN'
when a network failure on a node is detected.

To keep the test simple enough to verify the Neutron API, we don't
consider a scenario where multiple NIC ports could be present on that
particular node. It could still be done, though.

JIRA: DOCTOR-107

Change-Id: I9ffa0a23626b7edfaeb601dc90d55c756216cc41
Signed-off-by: Carlos Goncalves <carlos.goncalves@neclab.eu>