Delete project after stopping Inspector
At cleanup phase, the sample Inspector fails to re-enable the state of
the compute services. Such happens because the novaclient instance it's
using is bound to the 'doctor' project, which has just been deleted at
user.delete(). Thus, it is necessary to postpone the deletion of the
project after stopping the inspector.

Change-Id: I8cd5885cdfd3671a6c28342bdee65eac64dafaa8
Signed-off-by: Carlos Goncalves <carlos.goncalves@neclab.eu>