Update patch set 1
Patch Set 1:

Thanks, I still have not find working instructions to build the features container with patch under review in Doctor that change requirements.txt. This would be handy to get at some point as it is not the same to add manually updated packages to existing container as well as the changed Doctor code. This is more related to functest documentation, but also a thing for us. I am not expecting to have this now in this document, but nice to have in future.

Patch-set: 1
Label: Code-Review=0