Update patch set 7
Patch Set 7:

(10 comments)

Thanks for the comments Ifat. Have to explain some things better. One would currently understand some of these if have seen the presentation, but document is not good if it cannot explain things just by reading it. Still I want to remind that this is a guideline to be updated also in future and while it goes even to some very specific attributes, it will be in subject to change as the real implementation is just starting. Those attributes are still meaningful to tell what it takes to implement. This means, yes, need to fix some obvious bad naming used, but not like try to make this perfect.

Patch-set: 7