Update patch set 7
Patch Set 7:

I recall sometimes in the past it might have taken longer, but next time be in limits. I had the feeling if system was not used, the first API usage might have been longer. Also there have been issue with AODH not to be up and functional yet.

Patch-set: 7
Label: Verified=0