Update patch set 28
Patch Set 28:

I have now tested the latest changes 50735 + 53275 on my Apex env with 1 controller + 4 computes. This is not public env, so cannot make it as pod. If this is convincing enough, we could merge the code to quit continuous rebasing of these patch sets. Log for this run found here: https://pastebin.com/FBqCJRbh

Patch-set: 28