Update patch set 8
Patch Set 8:

Ifat will not be able to review this now and is ok if this gets merged. Whole document is anyhow subject to change later as this is more or less done against our current maintenance test case sample implementation. As OpenStack Fenix goes onward, we should update this according to better knowledge at that time (after Gambia)

Patch-set: 8