Update patch set 9
Patch Set 9:

Thanks Wenjuan, I have manually tried to run and there is still a bit more. One thing is that python3 is used to run those config changes on nodes, but it needs yaml that is not found for python3 if I do not install manually. Maybe get it after that.

Patch-set: 9