Update patch set 3
Patch Set 3: Code-Review+1

This should now be it. I think there is no way we get congress as a success today, so it will not change this. I can see there is a try to run functest, but fails before reaching features. That should be the only thing that could change this release note: https://build.opnfv.org/ci/view/functest/job/functest-apex-baremetal-daily-gambia/

Patch-set: 3
Label: Code-Review=+1