Update patch set 3
Patch Set 3:

ok, so the releng change was also not the best "all" should have remained in "sample" multijob. Have to see when you get the ZTE pod back up, how shall we do with the testing then. We might need Nokia POD to run "sample all" and "congress fault management" in case we do not get good results in virtual env.

Patch-set: 3