Update patch set 4
Patch Set 4:

Well, now it at least runs the test. Apex problem seems to be that it doesn't configure anything. Congress does not even work that bit that it did before because of this. Computes do not have nova.conf at all. One nice fact was also that even with sample inspector fault_management first worked nicely, but after a while, it never succeeds to be under one second (jumped 'permanently' from 300ms to 1300ms). Do not know if we saw the same with the virtual env we ramped down as we couldn't get good results. As always, the result is that we need to struggle to get even the basics working.

Patch-set: 4