Update patch set 7
Patch Set 7:

Some instability with testing environments and Congress surely does not work and not even trying it to this Friday release tag.
Nokia POD issues: Some python cache made as root, so jenkins user cannot clean up previous test. A too long lasting test may have ssh tunnel timeout. MCP seems to have variation in notification to consumer. Anyhow to me, we should merge this now and use it as base of testing in release note. I start working on that tomorrow, so get it on friday.

Patch-set: 7