Update patch set 10
Patch Set 10:

Great, works now. I do not think I will update this patch set further. I have developed some more features for VNFM in the Kubernetes case nad new workflow for that, but that should come later. Currently, that work is also only in Fenix space as it is a special deployment. I Can share it in the next meeting.

Patch-set: 10